backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.1
numpy==1.26.4
Authlib==1.3.0
httpx==0.27.0
itsdangerous==2.2.0
slowapi
pydantic[email]
python-decouple 